requirements.txt
openai==0.18.1
Pillow==9.5.0
streamlit==1.22.0
torch==2.0.1
tqdm==4.65.0
transformers==4.28.1
